+++ dms-20250414/alibabacloud_dms20250414/models/_list_data_agent_workspace_response_body.py
@@ -157,6 +157,7 @@ class ListDataAgentWorkspaceResponseBodyDataContent(DaraModel):
         create_time: int = None,
         creator: str = None,
         description: str = None,
+        is_session_share_enabled: bool = None,
         modify_time: int = None,
         role_name: str = None,
         total_member: int = None,
@@ -167,6 +168,7 @@ class ListDataAgentWorkspaceResponseBodyDataContent(DaraModel):
         self.create_time = create_time
         self.creator = creator
         self.description = description
+        self.is_session_share_enabled = is_session_share_enabled
         self.modify_time = modify_time
         self.role_name = role_name
         self.total_member = total_member
@@ -190,6 +192,9 @@ class ListDataAgentWorkspaceResponseBodyDataContent(DaraModel):
 
         if self.description is not None:
             result['Description'] = self.description
+
+        if self.is_session_share_enabled is not None:
+            result['IsSessionShareEnabled'] = self.is_session_share_enabled
 
         if self.modify_time is not None:
             result['ModifyTime'] = self.modify_time
@@ -221,6 +226,9 @@ class ListDataAgentWorkspaceResponseBodyDataContent(DaraModel):
 
         if m.get('Description') is not None:
             self.description = m.get('Description')
+
+        if m.get('IsSessionShareEnabled') is not None:
+            self.is_session_share_enabled = m.get('IsSessionShareEnabled')
 
         if m.get('ModifyTime') is not None:
             self.modify_time = m.get('ModifyTime')