+++ dms-20250414/alibabacloud_dms20250414/models/_update_data_agent_space_info_request.py
@@ -8,11 +8,13 @@ class UpdateDataAgentSpaceInfoRequest(DaraModel):
     def __init__(
         self,
         dmsunit: str = None,
+        is_session_share_enabled: bool = None,
         workspace_desc: str = None,
         workspace_id: str = None,
         workspace_name: str = None,
     ):
         self.dmsunit = dmsunit
+        self.is_session_share_enabled = is_session_share_enabled
         self.workspace_desc = workspace_desc
         self.workspace_id = workspace_id
         self.workspace_name = workspace_name
@@ -27,6 +29,9 @@ class UpdateDataAgentSpaceInfoRequest(DaraModel):
             result = _map
         if self.dmsunit is not None:
             result['DMSUnit'] = self.dmsunit
+
+        if self.is_session_share_enabled is not None:
+            result['IsSessionShareEnabled'] = self.is_session_share_enabled
 
         if self.workspace_desc is not None:
             result['WorkspaceDesc'] = self.workspace_desc
@@ -43,6 +48,9 @@ class UpdateDataAgentSpaceInfoRequest(DaraModel):
         m = m or dict()
         if m.get('DMSUnit') is not None:
             self.dmsunit = m.get('DMSUnit')
+
+        if m.get('IsSessionShareEnabled') is not None:
+            self.is_session_share_enabled = m.get('IsSessionShareEnabled')
 
         if m.get('WorkspaceDesc') is not None:
             self.workspace_desc = m.get('WorkspaceDesc')